Added test array; refactored uniquetol for 1D
Stored a test array of 729 floats in a helper file (the array consists of several distinct groups of very very similar elements with the minimum and maximum sufficiently far apart, all reshuffled around). Also refactored the uniquetol arguments from an is_highest boolean to an Occurrence enum for better clarity.