Added naive implementation of generalization to higher dims
Just a brute-force placeholder implementation (iterating over all k-combinations of the flattened subarrays for each k) while I think of a better one... This version is not particularly stable, and bug fixes/improvements will be happening in the ndarray branch. Just want to make whatever present functionality we have for ndarrays already available.